refactor: windowing 경로 변경
