+++ src/components/windowing/utils/calculateVisibleBounds.ts
